STOOPID

+++ Car.hpp
@@ -9,16 +9,16 @@ using namespace sf;
 class Car
 {
 private:
-    float angleInc = 3.141 / 3.0f;
+    double angleInc = 3.141 / 3.0f;
 public:
-    float heading;
-    array<float, 2> center;
-    array<array<float, 2>, 4> corners;
-    float size;
-    Car(float s, float h, array<float, 2> c);
-    void move(float d);
-    void rotate(float angle);
-    vector<float> getDists(Image im);
+    double heading;
+    array<double, 2> center;
+    array<array<double, 2>, 4> corners;
+    double size;
+    Car(double s, double h, array<double, 2> c);
+    void move(double d);
+    void rotate(double angle);
+    vector<double> getDists(Image im);
     array<array<double, 2>, 3> drawRay(Image im);
     ~Car();
 };
